Code review of pc02-graffiti-grad-iska9350
This project is a great example turtle graffiti. Isha fulfilled all the requirements of the assignment and used their turtle efficiently to draw comprehensive graphics, and some classic school notebook style doodles on his face. In order to improve upon this project, I added a for loop to draw the stop sign, and added two turtles named warning and bezos for graphics and doodling respectively. I also reorganized the variables and made edits to the order and categorization of the turtle drawing commands. I really liked the drawing here, and especially thought that Isha used the basic turtle commands extremely successfully to make both overlaid graphics and drawings relational to the image. My edits were mostly organizational, in order to strengthen the backend so that Isha or other collaborators can continue to build on their creation.